A one-byte undefined slot is one byte on every path that declares it
Ghidra types 1382 data symbols bare `undefined`, every one of them size 1.
extract/globals.ts mapped that to `auto`, and the two places that then had to
turn `auto` into a real declaration disagreed: normalizeGhidraType wrote
`void*` (4 bytes, and a pointer rather than an integer slot) while the
static-local and file-local paths wrote `int` (4 bytes). undefined1/2/4/8 all
keep their width through four mapping tables; only the bare spelling did not.

All three now give what undefined1 gives, uint8_t, so a symbol that is file-local
in one unit and extern in another is the same type on both sides - `&sym` was
`int*` against a `uint8_t*` reference wherever those two met.

+++ packages/reconstruct/src/codegen/impl.ts
@@ -1061,7 +1061,10 @@ export function generateImplementation(
       // dropped genuinely-used globals → "'x' was not declared". An emitted-but-
       // unused static is at worst a warning, never an error.
 
-      if (type === 'auto') type = 'int';
+      // `auto` is a one-byte `undefined` slot, and it must resolve to the SAME
+      // type globals.h gives the same symbol when it is not file-local - `int`
+      // here made `&sym` an `int*` against a `uint8_t*` reference.
+      if (type === 'auto') type = 'uint8_t';
 
       // A file-local static IS a declaration — and the only one this symbol
       // gets. The closure pass must not add an `extern` for it: that would be a
@@ -1098,7 +1101,7 @@ export function generateImplementation(
           }
         }
       } else {
-        if (type === 'auto') type = 'int';
+        if (type === 'auto') type = 'uint8_t';
         const arrayInfo = inferArrayDeclaration(global);
         if (arrayInfo) {
           fileLocalLines.push(`static ${arrayInfo.type} ${name}[${arrayInfo.count}];`);
